# shared/openapi/api.yaml
openapi: 3.0.3
info:
  title: User & Micropost API
  description: |
    ユーザーとマイクロポスト管理API (1:多の関係)
    - 読み取り専用設計（update/destroy操作なし）
    - 認証・認可なし（シンプル構成）
    - 高速プロトタイピング用途
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api/v1
    description: 開発環境
  - url: https://api.example.com/v1
    description: 本番環境
paths:
...
  /auth/me:
    get:
      summary: 自分のユーザー情報取得
      description: |-
        認証されたユーザーの詳細情報を取得します。
        トークンから自動的にユーザーを特定します。
      operationId: getCurrentUser
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ユーザー情報の取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
              examples:
                current_user:
                  summary: 現在のユーザー情報
                  value:
                    data:
                      id: 123
                      name: "田中太郎"
                      email: "tanaka@example.com"
                      bio: "フルスタック開発者です"
                      location: "東京"
                      website: "https://tanaka.dev"
                      roles: ["user", "readonly-admin"]
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-02-10T16:45:00Z"
                      micropostCount: 42
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/password:
    put:
      summary: パスワード変更
      description: |-
        認証されたユーザーのパスワードを変更します。
        現在のパスワードの確認が必要です。
      operationId: changePassword
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            examples:
              change_password_example:
                summary: パスワード変更の例
                value:
                  currentPassword: "OldPassword123"
                  newPassword: "NewSecurePass456"
      responses:
        '200':
          description: パスワード変更成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - success
                      - message
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "パスワードが正常に変更されました"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/forgot-password:
    post:
      summary: パスワードリセット要求
      description: |-
        パスワードリセット用のメールを送信します。
        セキュリティのため、存在しないメールアドレスでも成功レスポンスを返します。
      operationId: forgotPassword
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
            examples:
              forgot_password_example:
                summary: パスワードリセット要求の例
                value:
                  email: "user@example.com"
      responses:
        '200':
          description: リセットメール送信成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - success
                      - message
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "パスワードリセットメールを送信しました（該当するアカウントが存在する場合）"
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/reset-password:
    post:
      summary: パスワードリセット実行
      description: |-
        リセットトークンを使用してパスワードを新しく設定します。
        成功時は全デバイスからログアウトされます。
      operationId: resetPassword
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
            examples:
              reset_password_example:
                summary: パスワードリセットの例
                value:
                  token: "reset-token-from-email"
                  newPassword: "NewSecurePass789"
      responses:
        '200':
          description: パスワードリセット成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    required:
                      - success
                      - message
                    properties:
                      success:
                        type: boolean
                        example: true
                      message:
                        type: string
                        example: "パスワードが正常にリセットされました"
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      summary: アクセストークンリフレッシュ
      description: |-
        リフレッシュトークンを使用して新しいアクセストークンを取得します。
        Keycloak互換のインターフェースを提供します。
      operationId: refreshToken
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              refresh_token_example:
                summary: トークンリフレッシュの例
                value:
                  refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: トークンリフレッシュ成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                refresh_success:
                  summary: リフレッシュ成功レスポンス
                  value:
                    data:
                      user:
                        id: 123
                        name: "田中太郎"
                        email: "tanaka@example.com"
                        roles: ["user"]
                        createdAt: "2024-01-15T10:30:00Z"
                      tokens:
                        access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        token_type: "Bearer"
                        expires_in: 3600
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: リフレッシュトークンが無効です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_refresh_token:
                  summary: 無効なリフレッシュトークン
                  value:
                    error: "invalid_refresh_token"
                    message: "リフレッシュトークンが無効です"
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ========== ユーザー関連エンドポイント ==========
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT認証トークン。
        Keycloak互換形式でaccess_tokenを使用。

  # ========== パラメータ定義 ==========
...
  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            bad_request:
              summary: 不正なリクエスト
              value:
                error: "bad_request"
                message: "リクエストパラメータが不正です"
                details:
                  - "page パラメータは1以上の整数である必要があります"
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: 認証エラー
              value:
                error: "unauthorized"
                message: "有効な認証トークンが必要です"
...
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: バリデーションエラー
              value:
                error: "validation_error"
                message: "入力データが不正です"
                details:
                  - "content: 1文字以上280文字以下で入力してください"
                  - "content: 必須フィールドです"
    InternalServerError:
      description: サーバー内部エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: サーバーエラー
              value:
                error: "internal_server_error"
                message: "サーバー内部でエラーが発生しました"

  # ========== スキーマ定義 ==========
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: エラーコード
          example: "validation_error"
        message:
          type: string
          description: エラーメッセージ（日本語）
          example: "入力データが不正です"
        details:
          type: array
          description: 詳細なエラー情報（バリデーションエラー等）
          items:
            type: string
          example:
            - "content: 1文字以上280文字以下で入力してください"
        requestId:
          type: string
          description: リクエストID（デバッグ用）
          example: "req_12345"
        timestamp:
          type: string
          format: date-time
          description: エラー発生日時
          example: "2024-02-10T16:45:00Z"

    # ========== 認証関連スキーマ ==========
...
    AuthTokens:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWTアクセストークン
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          enum: [Bearer]
          description: トークンタイプ
          example: "Bearer"
        expires_in:
          type: integer
          minimum: 1
          description: トークンの有効期限（秒）
          example: 3600
    AuthUser:
      type: object
      required:
        - id
        - name
        - email
        - roles
        - createdAt
      properties:
        id:
          type: integer
          minimum: 1
          description: ユーザーID
          example: 123
        name:
          type: string
          description: ユーザー名
          example: "田中太郎"
        email:
          type: string
          format: email
          description: メールアドレス
          example: "tanaka@example.com"
        bio:
          type: string
          maxLength: 200
          description: 自己紹介文
          example: "フルスタック開発者です"
        location:
          type: string
          maxLength: 50
          description: 居住地
          example: "東京"
        website:
          type: string
          format: uri
          maxLength: 200
          description: ウェブサイトURL
          example: "https://tanaka.dev"
        roles:
          type: array
          items:
            type: string
            enum: [user, readonly-admin, admin]
          description: ユーザーの権限ロール
          example: ["user"]
        createdAt:
          type: string
          format: date-time
          description: アカウント作成日時
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: 最終更新日時
          example: "2024-02-10T16:45:00Z"
        micropostCount:
          type: integer
          minimum: 0
          description: 投稿数
          example: 42
    AuthResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - user
            - tokens
          properties:
            user:
              $ref: '#/components/schemas/AuthUser'
            tokens:
              $ref: '#/components/schemas/AuthTokens'
    CurrentUserResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/AuthUser'

    # ========== パスワード管理スキーマ ==========
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: 現在のパスワード
          example: "OldPassword123"
        newPassword:
          type: string
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[a-zA-Z])(?=.*\d).{8,}$'
          description: 新しいパスワード（8文字以上、英数字を含む）
          example: "NewSecurePass456"
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: パスワードリセット用メールアドレス
          example: "user@example.com"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          description: パスワードリセットトークン（メールから取得）
          example: "reset-token-from-email"
        newPassword:
          type: string
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[a-zA-Z])(?=.*\d).{8,}$'
          description: 新しいパスワード（8文字以上、英数字を含む）
          example: "NewSecurePass789"
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: リフレッシュトークン
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    # ========== Keycloak互換スキーマ ==========
...